requirements.txt
google-adk
google-genai
trafilatura
python-dotenv
streamlit